Fixed bug 3065 - Screen is flickering during closing on-screen keyboard on Android
Deve

When I'm trying to close on-screen keyboard using SDL_StopTextInput() function by touching the screen (SDL_FINGERUP or SDL_FINGERDOWN event), the screen is flickering. It is white for a while.

Note that it usually works without problems when I use phone's "back" button. I noticed flickering occasionally too, but not that often.

Philipp Wiesemann

The attached patch maybe fixes the flicker but not the actual fault causing it.